Added lip and slop dimensions
This change allows the worksheet to properly calculate the lengths to
use when cutting the parts using the picture frame cut jig. Before this
change, if you set the jig to the _Inner Length_ of the cutlist, the
frame would end up being about 1/4" too small. The inside cut length
needs to be longer to allow for the lip and slop to be taken into
account. The values default to 1/8", which should be reasonable. If
they're too long, its easier to cut the parts shorter than it is to
make them longer :p